refactor: Banner의 구조 개선
