+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/Banner/external/GPTService.java
@@ -1,4 +1,4 @@
-package com.techeerpicture.TecheerPicture.Banner;
+package com.techeerpicture.TecheerPicture.Banner.external;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
@@ -10,6 +10,7 @@ import org.springframework.web.client.RestTemplate;
 import java.util.List;
 import java.util.Map;
 
+import com.techeerpicture.TecheerPicture.Banner.util.GeneratedTexts;
 /**
  * GPTService
  *
