+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/Banner/util/GeneratedTexts.java
@@ -1,4 +1,4 @@
-package com.techeerpicture.TecheerPicture.Banner;
+package com.techeerpicture.TecheerPicture.Banner.util;
 
 public class GeneratedTexts {
